+++ Plugin/PluginTwitch/Positioned.cs
@@ -8,9 +8,9 @@ namespace PluginTwitchChat
 {
     public interface Positioned
     {
-        float X { get; set; }
-        float Y { get; set; }
-        float Width { get; set; }
-        float Height { get; set; }
+        double X { get; set; }
+        double Y { get; set; }
+        double Width { get; set; }
+        double Height { get; set; }
     }
 }
